requirements.txt
notion-client
playwright
playwright-stealth
python-dotenv
httpx
mcp
